[Spec Decode] Move draft_gumbel_pos to spec_decode utils
